+++ app/friend/search/page.tsx
@@ -5,7 +5,7 @@ import { useDebounce } from "@/hooks/useDebounce";
 import SearchInput from "@/components/friend/search/SearchInput";
 import BasicTitle from "@/components/friend/BasicTitle";
 import SearchResultList from "@/components/friend/search/SearchResultList";
-import { searchResultList } from "@/app/mocks/searchResults";
+import { searchResultList } from "@/mocks/searchResults";
 
 export default function SearchPage() {
   const [query, setQuery] = useState("");
